backend/pyproject.toml
[tool.poetry]
name = "recipes"
version = "0.1.0"
description = ""
authors = ["Kasia"]

[tool.poetry.dependencies]
python = "^3.9"

Django = "^3.1.7"
django-environ = "^0.4.5"
django-extensions = "^3.1.0"
djangorestframework = "^3.12.4"

psycopg2-binary = "^2.8.6"
argon2-cffi = "^20.1.0"

gunicorn = "^20.0.4"
Collectfast = "^2.2.0"
sentry-sdk = "^1.0.0"

[tool.poetry.dev-dependencies]
Werkzeug = "^1.0.1"

coverage = "^5.5"
factory-boy = "^3.1.0"
mypy = "^0.770"
pytest = "^6.1.2"
pytest-cov = "^2.11.1"
pytest-deadfixtures = "^2.2.1"
pytest-django = "^4.1.0"
pytest-factoryboy = "^2.0.3"
pytest-isort = "^1.3.0"
pytest-mock = "^3.4.0"
pytest-randomly = "^3.5.0"
pytest-sugar = "^0.9.4"
pytest-timeout = "^1.4.2"
requests-mock = "^1.8.0"

flake8 = "^3.9.0"
flake8-django = "^1.1.1"
flake8-pytest = "^1.3"
flake8-pytest-style = "^1.3.0"
wemake-python-styleguide = "^0.14.1"

isort = "^4.3.21"

django-debug-toolbar = "^3.2"
django-test-migrations = "^1.1.0"
django-coverage-plugin = "^1.6.0"
drf-spectacular = "^0.15.1"

safety = "^1.10.3"
django-silk = "^4.1.0"
django-stubs = "^1.5.0"
